requirements.txt
Django==4.2.27
djangorestframework==3.16.1
celery==5.6.0
redis==7.1.0
psycopg2-binary==2.9.11
drf-spectacular==0.29.0
gunicorn==21.2.0
